MLP essential files
the 2 python files with "MLP" in their names uses the same MLP technique, however, the MLP.py uses binary classification method, where there is only 1 output \belongs [0, 1], and the classification is determined by the size of that point. The two_outputs_MLP.py, on the other hand, have 2 outputs, and the classification will be determined based on the difference between the 2 outputs. In principle, these two method should have the same performance. In the report, results are presented using the two_outputs_MLP.py.

the txt file store the accuracies and the scores associated to the two_outputs_MLP, for details, please refer to the report.

the ipynb file (jupyter notebook) is used for the generation of all graphs used in the report.